+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.7.0
+
+- Added MemoryStorageConnector (useful for testing)
+- Added `load` method to `CloudWandererResource`
+- Added `load` support to `MemoryStorageConnector`
+- Added `load` support to `DynamoDbConnector`
+
 # 0.6.0
 
 - Fetch region information for semi-global resources like S3 buckets using `GlobalServiceMapping` objects.